requirements.txt
marshmallow==2.15.0
nose==1.3.7
pkg-resources==0.0.0
PyMySQL==0.8.0
SQLAlchemy==1.1.15
Flask-RESTful==0.3.6
PyYAML==3.12
